moves signIn/signUp form into pages as theyre isolated
