+++ src/pages/SigninForm.js
@@ -1,37 +1,38 @@
 import { Button, InputLabel, TextField, Typography } from "@mui/material";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { signUp } from "../services/authServices";
+import { signIn } from "../services/authServices";
 import { useGlobalState } from "../context/stateContext";
-import { ErrorAlert } from "./ErrorAlert";
+import { ErrorAlert } from "../components/ErrorAlert";
 
-const SignupForm = () => {
+const SigninForm = () => {
   const { dispatch } = useGlobalState();
   const navigate = useNavigate();
 
   const initialFormData = {
-    firstName: "",
-    lastName: "",
     username: "",
-    email: "",
     password: "",
-    password_confirmation: "",
   };
   const [formData, setFormData] = useState(initialFormData);
   const [error, setError] = useState();
 
   const handleSubmit = (e) => {
     e.preventDefault();
+    setError(null);
 
     try {
       if (Object.values(formData).includes("")) {
         throw Error("Fields cannot be empty");
       }
 
-      signUp(formData)
+      signIn(formData)
         .then((user) => {
+          setError();
+          console.log("THIS IS USER", user);
           sessionStorage.setItem("username", user.username);
           sessionStorage.setItem("token", user.jwt);
+          sessionStorage.setItem("firstName", user.firstName);
+          sessionStorage.setItem("lastName", user.lastName);
           dispatch({
             type: "setLoggedInUser",
             data: user.username,
@@ -40,11 +41,14 @@ const SignupForm = () => {
             type: "setToken",
             data: user.jwt,
           });
+          dispatch({
+            type: "setUserDetails",
+            data: { firstName: user.firstName, lastName: user.lastName },
+          });
           setFormData(initialFormData);
           navigate("/");
         })
         .catch((err) => {
-          console.log(err);
           setError(err);
         });
     } catch (err) {
@@ -58,47 +62,20 @@ const SignupForm = () => {
       [e.target.id]: e.target.value,
     });
   };
-
   return (
     <>
       {error && <ErrorAlert message={error.message} setError={setError} />}
+
       <div className="form-container">
         <form onSubmit={handleSubmit}>
-          <Typography variant="h4">Register user</Typography>
+          <Typography variant="h4">Sign in</Typography>
 
-          <InputLabel>First Name:</InputLabel>
-          <TextField
-            type="text"
-            name="firstName"
-            id="firstName"
-            value={formData.firstName}
-            onChange={handleFormData}
-          />
-
-          <InputLabel>Last Name:</InputLabel>
-          <TextField
-            type="text"
-            name="lastName"
-            id="lastName"
-            value={formData.lastName}
-            onChange={handleFormData}
-          />
-
-          <InputLabel>Username:</InputLabel>
+          <InputLabel>Username / Email:</InputLabel>
           <TextField
             type="text"
             name="username"
             id="username"
             value={formData.username}
-            onChange={handleFormData}
-          />
-
-          <InputLabel>Email:</InputLabel>
-          <TextField
-            type="text"
-            name="email"
-            id="email"
-            value={formData.email}
             onChange={handleFormData}
           />
 
@@ -111,17 +88,8 @@ const SignupForm = () => {
             onChange={handleFormData}
           />
 
-          <InputLabel htmlFor="password">Password confirmation:</InputLabel>
-          <TextField
-            type="password"
-            name="password_confirmation"
-            id="password_confirmation"
-            value={formData.password_confirmation}
-            onChange={handleFormData}
-          />
-
           <Button variant="contained" type="submit">
-            Sign up
+            Login
           </Button>
         </form>
       </div>
@@ -129,4 +97,4 @@ const SignupForm = () => {
   );
 };
 
-export default SignupForm;
+export default SigninForm;
